requirements.txt
fastapi
uvicorn
pydantic
google-auth
google-auth-oauthlib
google-api-python-client
openai
pydantic
typing
requests==2.32.5
python-dotenv